+++ src/components/common/Modal/BottomShareModal/BottomShareModal.module.scss
@@ -1,0 +1,52 @@
+@layer components {
+  .modalBase {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    background-color: #ffffff;
+    &[data-type='bottom'] {
+      @include modal-base;
+    }
+    &[data-type='share'] {
+      @include share-modal-base;
+    }
+    z-index: $z-index-modal;
+  }
+
+  .modalContent {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    background-color: #ffffff;
+    &[data-type='bottom'] {
+      padding: 6rem 4rem;
+      border-radius: 1.2rem 1.2rem 0rem 0rem;
+      box-shadow: 0rem 0.4rem 1.9rem 0rem rgba(0, 0, 0, 0.25);
+    }
+    &[data-type='share'] {
+      padding: 4rem 2rem 3.2rem;
+      border-radius: 1.6rem;
+      box-shadow: 0rem 0.4rem 2rem 0rem rgba(0, 0, 0, 0.08);
+      width: 23rem;
+      height: 34.1rem;
+    }
+  }
+
+  .modalWrapper {
+    width: 100%;
+    height: 100%;
+  }
+
+  .modalBackground {
+    position: absolute;
+    top: 0;
+    bottom: 0;
+    left: 0;
+    right: 0;
+    &[data-type='share'] {
+      background-color: rgba(0, 0, 0, 0.2);
+    }
+  }
+}
