+++ src/components/common/Modal/BottomShareModal/index.tsx
@@ -1,15 +1,16 @@
 import classNames from 'classnames/bind';
 import { useModalProps } from '@/hooks/useModal';
-import ModalBase, { ModalProps } from '@/components/common/Modal/BottomModal/ModalBase';
-import styles from './BottomModal.module.scss';
+import ModalBase, { ModalProps } from '@/components/common/Modal/BottomShareModal/ModalBase';
+import styles from './BottomShareModal.module.scss';
 import { PropsWithChildren } from 'react';
 
 const cx = classNames.bind(styles);
 
-export default function BottomModal({
+export default function BottomShareModal({
   modalOpen,
   handleModalOpen,
   handleModalClose,
+  type,
   className,
   children,
 }: PropsWithChildren<useModalProps & ModalProps>) {
@@ -22,7 +23,7 @@ export default function BottomModal({
         }
       }}>
       {modalOpen && (
-        <ModalBase className={className} onClose={handleModalClose}>
+        <ModalBase type={type} className={className} onClose={handleModalClose}>
           {children}
         </ModalBase>
       )}
@@ -33,18 +34,18 @@ export default function BottomModal({
 {
   /* 사용법
 1. 사용처에서 import useModal from '@/hooks/useModal';
-2. 사용처에서 import BottomModal from '@/components/common/BottomModal';
+2. 사용처에서 import BottomShareModal from '@/components/common/BottomShareModal';
 
   const { modalOpen, handleModalOpen, handleModalClose } = useModal();
   return (
     <>
-            <BottomModal
+            <BottomShareModal
               className={cx('')}
               modalOpen={modalOpen}
               handleModalOpen={handleModalOpen}
               handleModalClose={handleModalClose}>
               *모달 내부 디자인 추가*
-            </BottomModal>
+            </BottomShareModal>
     </>
   );
   사용 예시는 pages > text > my.tsx에서 확인할 수 있습니다.
